@@ -6,9 +6,10 @@
     "start": "next start"
   },
   "dependencies": {
+    "dependencies": "^0.0.1",
     "next": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@types/node": "^18.11.9",
