@@ -1,18 +1,23 @@
-import { useRouter } from "next/router";
+"use client"; // Required for useParams() and accessing document.cookie
+
+import { useParams } from "next/navigation";
 
 function hasMiddlewareMatched(slug?: string[]) {
+  if (typeof document === "undefined") return false;
+
   const values =
-    (typeof document !== "undefined" ? document.cookie : "")
+    document.cookie
       .split(";")
       .map((pair) => pair.split("="))
       .filter(([key]) => key === "middleware-slug")
       .map(([, value]) => value.trim()) ?? [];
+
   return values.some((value) => value === slug?.join("/"));
 }
 
 export default function ContentPage() {
-  const router = useRouter();
-  const slug = router.query.slug as string[]; // slug is an array of path segments
+  const params = useParams();
+  const slug = Array.isArray(params.slug) ? params.slug : []; // Ensure it's always an array
 
   return (
     <>
